add config files
修改ffmpeg目录名字，用于include
修改引用
添加include在某些引用头文件中

+++ ga/module/encoder-video/encoder-video.cpp
@@ -178,12 +178,12 @@ vencoder_reconfigure(int iid) {
 		for (int i = 0; i < rtspconf->vso->size(); i += 2) {
 			if ((*rtspconf->vso)[i].compare("b") == 0) {
 				if (reconf->bitrateKbps > 0)
-					(*rtspconf->vso)[i+1] = std::to_string(reconf->bitrateKbps * 1000);
+					(*rtspconf->vso)[i+1] = std::to_string((long double)reconf->bitrateKbps * 1000);
 				continue;
 			}
 			if ((*rtspconf->vso)[i].compare("crf") == 0) {
 				if (reconf->crf > 0)
-					(*rtspconf->vso)[i+1] = std::to_string(reconf->crf);
+					(*rtspconf->vso)[i+1] = std::to_string((long double)reconf->crf);
 				continue;
 			}
 		}
